@@ -4,12 +4,13 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "doctrine/doctrine-bundle": "^2.5",
+        "doctrine/annotations": "*",
+        "doctrine/cache": "*",
         "doctrine/doctrine-migrations-bundle": "^3.2",
-        "doctrine/orm": "*",
+        "doctrine/orm": "^2.11",
         "sensio/framework-extra-bundle": "^6.2",
         "symfony/asset": "5.4.*",
         "symfony/console": "5.4.*",
